+++ src/gamspy/_symbols/set.py
@@ -24,12 +24,14 @@
 #
 from __future__ import annotations
 
+import itertools
 from typing import Any
 from typing import Literal
 from typing import TYPE_CHECKING
 
 import gams.transfer as gt
 import pandas as pd
+from gams.core.gdx import GMS_DT_SET
 
 import gamspy as gp
 import gamspy._algebra.condition as condition
@@ -325,6 +327,63 @@ class Set(gt.Set, operable.Operable, Symbol, SetMixin):
 
     """
 
+    @classmethod
+    def _constructor_bypass(
+        cls,
+        container: Container,
+        name: str,
+        domain: list[str | Set] = [],
+        is_singleton: bool = False,
+        records: Any | None = None,
+        description: str = "",
+    ):
+        # create new symbol object
+        obj = Set.__new__(
+            cls,
+            container,
+            name,
+            domain,
+            is_singleton,
+            records,
+            description=description,
+        )
+
+        # set private properties directly
+        obj._requires_state_check = False
+        obj._container = container
+        container._requires_state_check = True
+        obj._name = name
+        obj._domain = domain
+        obj._domain_forwarding = False
+        obj._description = description
+
+        obj._records = records
+        obj._modified = True
+        obj._is_singleton = is_singleton
+
+        # typing
+        if obj.is_singleton:
+            obj._gams_type = GMS_DT_SET
+            obj._gams_subtype = 1
+        else:
+            obj._gams_type = GMS_DT_SET
+            obj._gams_subtype = 0
+
+        # add to container
+        container.data.update({name: obj})
+
+        # gamspy attributes
+        obj._is_dirty = False
+        obj.where = condition.Condition(obj)
+        obj.container._add_statement(obj)
+        obj._current_index = 0
+
+        # miro support
+        obj._is_miro_input = False
+        obj._is_miro_output = False
+
+        return obj
+
     def __new__(
         cls,
         container: Container,
@@ -372,34 +431,97 @@ class Set(gt.Set, operable.Operable, Symbol, SetMixin):
         is_miro_input: bool = False,
         is_miro_output: bool = False,
     ):
-        self._is_dirty = False
-        self.where = condition.Condition(self)
-        name = validation.validate_name(name)
-
-        singleton_check(is_singleton, records)
-
-        super().__init__(
-            container,
-            name,
-            domain,
-            is_singleton,
-            records,
-            domain_forwarding,
-            description,
-            uels_on_axes,
-        )
-
-        validation.validate_container(self, self.domain)
-        self.container._add_statement(self)
-        self._current_index = 0
         self._is_miro_input = is_miro_input
         self._is_miro_output = is_miro_output
 
-        if is_miro_input:
-            self.container._miro_input_symbols.append(self.name)
+        # domain handling
+        if domain is None:
+            domain = ["*"]
 
-        if is_miro_output:
-            self.container._miro_output_symbols.append(self.name)
+        if isinstance(domain, (gp.Set, gp.Alias, str)):
+            domain = [domain]
+
+        # does symbol exist
+        has_symbol = False
+        if isinstance(getattr(self, "container", None), gp.Container):
+            has_symbol = True
+
+        if has_symbol:
+            try:
+                if not isinstance(self, Set):
+                    raise TypeError(
+                        f"Cannot overwrite symbol {self.name} in container"
+                        " because it is not a Set object)"
+                    )
+
+                if any(
+                    d1 != d2
+                    for d1, d2 in itertools.zip_longest(self.domain, domain)
+                ):
+                    raise ValueError(
+                        "Cannot overwrite symbol in container unless symbol"
+                        " domains are equal"
+                    )
+
+                if self.is_singleton != is_singleton:
+                    raise ValueError(
+                        "Cannot overwrite symbol in container unless"
+                        " 'is_singleton' is left unchanged"
+                    )
+
+                if self.domain_forwarding != domain_forwarding:
+                    raise ValueError(
+                        "Cannot overwrite symbol in container unless"
+                        " 'domain_forwarding' is left unchanged"
+                    )
+
+            except ValueError as err:
+                raise ValueError(err)
+
+            except TypeError as err:
+                raise TypeError(err)
+
+            # reset some properties
+            self._requires_state_check = True
+            self.container._requires_state_check = True
+            if description != "":
+                self.description = description
+            self.records = None
+            self.modified = True
+
+            # only set records if records are provided
+            if records is not None:
+                self.setRecords(records, uels_on_axes=uels_on_axes)
+
+        else:
+            self._is_dirty = False
+            self.where = condition.Condition(self)
+            name = validation.validate_name(name)
+
+            singleton_check(is_singleton, records)
+
+            super().__init__(
+                container,
+                name,
+                domain,
+                is_singleton,
+                records,
+                domain_forwarding,
+                description,
+                uels_on_axes,
+            )
+
+            if is_miro_input:
+                container._miro_input_symbols.append(self.name)
+
+            if is_miro_output:
+                container._miro_output_symbols.append(self.name)
+
+            validation.validate_container(self, self.domain)
+            self.container._add_statement(self)
+            self._current_index = 0
+
+            self.container._run()
 
     def __len__(self):
         if self.records is not None:
@@ -426,9 +548,7 @@ class Set(gt.Set, operable.Operable, Symbol, SetMixin):
         return self
 
     def __getitem__(self, indices: tuple | str) -> implicits.ImplicitSet:
-        domain = validation._transform_given_indices(self.domain, indices)
-
-        validation.validate_domain(self, domain)
+        domain = validation.validate_domain(self, indices)
 
         return implicits.ImplicitSet(self, name=self.name, domain=domain)
 
@@ -437,9 +557,7 @@ class Set(gt.Set, operable.Operable, Symbol, SetMixin):
         indices: tuple | str,
         assignment,
     ):
-        domain = validation._transform_given_indices(self.domain, indices)
-        validation.validate_container(self, domain)
-        validation.validate_domain(self, domain)
+        domain = validation.validate_domain(self, indices)
 
         if isinstance(assignment, bool):
             assignment = "yes" if assignment is True else "no"  # type: ignore
@@ -505,6 +623,10 @@ class Set(gt.Set, operable.Operable, Symbol, SetMixin):
                 # reset state check flags for all symbols in the container
                 for symbol in self.container.data.values():
                     symbol._requires_state_check = True
+
+    def setRecords(self, records: Any, uels_on_axes: bool = False) -> None:
+        super().setRecords(records, uels_on_axes)
+        self.container._run()
 
     def gamsRepr(self) -> str:
         """